% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relmatGlmer.R
\name{relmatGlmer}
\alias{relmatGlmer}
\title{Fit a generalized linear mixed model (GLMM) with relative matrices.}
\usage{
relmatGlmer(...)
}
\description{
Its implementation follows that of `glmer` function in lme4 R package.
}
