% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relmatLmer.R
\name{relmatLmer}
\alias{relmatLmer}
\title{Fit a linear mixed model (LMM) with relative matrices.}
\usage{
relmatLmer(...)
}
\description{
Its implementation follows that of lmer function in lme4 R package.
}
